Break out tests into small groups.
The way we are testing this in general is strange, due to the fact that
we are testing tests.  Meta.  Anyway, rather than testing a bunch of
tests at once, I think it makes more sense, and is easier to maintain,
if the tests are broken out as much as possible.